Add discrete variability in FMI 3 Feedthrough FMU
The model variables "Float32_discrete_input" and
"Float32_discrete_output" of the FMI 3 Feedthrough FMU are missing the
discrete variability attribute. This may cause issues in a standard
conforming importer because it may try to set the variable during an
illegal mode, e.g., during the Continuous Time mode.

This adds the discrete attribute to both variables which is in line with
the "Float64_discrete_input" and "Float64_discrete_output" variables
which already have the discrete variability attribute.